chore: splits out relationship
